Create new component for diseaes page table, move column show/hide logic there
