+++ webapp/src/containers/diseasePage/index.js
@@ -1,22 +1,7 @@
 import React, { Component } from 'react';
 import HeadMetaTags from '../../components/headMetaTags';
 import Subsection from '../../components/subsection';
-import { RemoteDataTable } from '../../components/dataTable';
-
-const columns = [
-  {
-    field: 'do_name',
-    label: 'Disease Name',
-    sortable: true,
-  },
-  {
-    field: 'do_id',
-    label: 'DO ID',
-    isKey: true,
-    sortable: true,
-    format: (id) => id + '!'
-  }
-];
+import { DiseasePageAssociationsTable } from '../../components/disease';
 
 class DiseasePage extends Component {
   render() {
@@ -36,7 +21,7 @@ class DiseasePage extends Component {
         </a>
 
         <Subsection hardcoded title='Associations'>
-          <RemoteDataTable columns={columns} url='http://localhost:3000/diseases' />
+          <DiseasePageAssociationsTable />
         </Subsection>
       </div>
     );
